Functions| Pointers | Week 4
Q1 : Write a function to print squares of first n natural numbers, taking n as argument to the function
Q2 : Write a function that takes the radius of a circle as an argument and returns its area.
Q3: Given two numbers a and b, write a function to print all odd numbers between them.
Q4: Write a function to count the number of digits in a number and then print the square of this number.
Q5: The minimum number of functions present in any C++ program is:

1
Q6: State True and False:
A function may be called more than once from any other function         yes
It is necessary for a function to return some value.              no
Q7: Explore:
Can the same function name be used for different functions without any conflict?         yes